% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-settings.R
\name{get_settings}
\alias{get_settings}
\title{Get Application Settings}
\usage{
get_settings()
}
\value{
A named list of settings
}
\description{
To ensure consistency between modules, this function returns a list of
settings for things like thresholds, colour schemes, etc.
}
