% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data From CloudStor}
\usage{
load_data(path)
}
\arguments{
\item{path}{The location of the dataset on CloudStor}
}
\value{
The loaded dataset
}
\description{
Load the dataset from CloudStor
}
