% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_data.R
\name{make_dummy_data}
\alias{make_dummy_data}
\title{Create dummy data}
\usage{
make_dummy_data(n_studies)
}
\arguments{
\item{n_studies}{number of studies to simulate.}
}
\value{
A list with the created dataset, and the exposure/outcome keys.
}
\description{
Creates a dummy dataset for testing and development.
}
\details{
The function takes the existing data and uses it as the basis of the dummy
studies. This means that studies that are created can only have
exposure/outcome combinations that are already in the original data.

In addition, the function returns two additional dataframes that contain the
different levels for outcomes and exposures.
}
