% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmap.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make Heatmap}
\usage{
make_heatmap(df, subcategories = FALSE, show = "Counts")
}
\arguments{
\item{df}{The tidy dataframe.}

\item{subcategories}{Show subcategories?.}

\item{show}{Units to show as cells in the heatmap.}
}
\value{
A ggplot object
}
\description{
Create the heatmap to descripve the exposure/outcome relationships.
}
