% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview_server}
\alias{overview_server}
\title{Overview Module Server}
\usage{
overview_server(id, dataset)
}
\arguments{
\item{id}{Namespace ID.}

\item{dataset}{The clean dataset.}
}
\value{
A Shiny UI module.
}
\description{
Overview Module Server
}
