% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\alias{tidy_data_effects}
\alias{tidy_data_reviews}
\alias{tidy_data_studies}
\alias{tidy_data_rob}
\title{Tidy Data}
\usage{
tidy_data_effects(df)

tidy_data_reviews(df)

tidy_data_studies(df)

tidy_data_rob(df)
}
\arguments{
\item{df}{The dataset to tidy.}
}
\value{
A tidy dataset
}
\description{
Tidies the dataset for visualisation.
}
