% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mod_heatmap.R
\name{update_curr}
\alias{update_curr}
\title{Update the current list of exposures or outcomes.}
\usage{
update_curr(key, types, level, clicked, clicked_level)
}
\arguments{
\item{key}{The exposures or outcomes 'key'. \code{\link{make_dummy_data}}
provides an example of the expected key format.}

\item{types}{A list of exposure types, where the names match the column
names of the key.}

\item{level}{The current level, as an int.}

\item{clicked}{The value that was clicked, or \code{NULL} if nothing was
clicked.}

\item{clicked_level}{The level the click occured, where applicable.}
}
\value{
A list of exposures or outcomes.
}
\description{
Takes information about the current events and uses it to determine the
next list of exposures or outcomes that should be used.
}
\keyword{internal}
