% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mod_heatmap.R
\name{update_level}
\alias{update_level}
\title{Update level}
\usage{
update_level(types, level, clicked)
}
\arguments{
\item{types}{The list of possible types the exposure or outcome can take.}

\item{level}{The current level as an int.}

\item{clicked}{The value that was clicked, or NULL if nothing was clicked.}
}
\value{
The updated level value as an int.
}
\description{
Update the level of the current exposure and outcome.
}
\keyword{internal}
