% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data From CloudStor}
\usage{
load_data(path, use_cache = TRUE)
}
\arguments{
\item{path}{The location of the dataset on CloudStor}

\item{use_cache}{Use a cached version of the dataset, if available?}
}
\value{
The loaded dataset
}
\description{
Load the dataset from CloudStor
}
